Organise ML model source
